fix: screen shot

+++ apps/mobile/src/core/services/browserService.ts
@@ -375,8 +375,9 @@ export class BrowserService extends StoreServiceBase<BrowserStore, 'browser'> {
   }
 
   async saveScreenshot({ tempUri, tabId }: { tempUri: string; tabId: string }) {
-    const fileName = `screenshot-${tabId}.jpg`;
+    const fileName = `screenshot-${tabId}-${Date.now()}.jpg`;
     const filePath = `${RNFS.DocumentDirectoryPath}/${fileName}`;
+    await this.removeScreenshot({ tabId });
     try {
       if (await RNFS.exists(filePath)) {
         await RNFS.unlink(filePath);
@@ -389,8 +390,11 @@ export class BrowserService extends StoreServiceBase<BrowserStore, 'browser'> {
   }
 
   async removeScreenshot({ tabId }: { tabId: string }) {
-    const fileName = `screenshot-${tabId}.jpg`;
-    const filePath = `${RNFS.DocumentDirectoryPath}/${fileName}`;
+    const tab = this.store.browserTabs.tabs.find(item => item.id === tabId);
+    const filePath = tab?.viewShot;
+    if (!filePath) {
+      return;
+    }
     try {
       if (await RNFS.exists(filePath)) {
         await RNFS.unlink(filePath);
